Cleaned up processes
