+++ src/sys/process.c
@@ -1,36 +1,28 @@
 #include "sys/process.h"
-#include "drivers/console.h"
 #include "drivers/printk.h"
+#include "lib/stdlib.h"
+#include "lib/string.h"
 #include "memory/consts.h"
 #include "memory/kmalloc.h"
 #include "memory/memory.h"
 
+#include <stdbool.h>
 #include <stddef.h>
 #include <stdint.h>
-#include <string.h>
 
 extern uint32_t page_directory[1024];
-extern proc_t *current_proc;
 
 proc_t ptables[NUM_PROC];
+proc_t *current_proc = NULL;
+context_t *scheduler_context;
 uint32_t pid_counter = 1;
 
-void process_list(void) {
-  printk("PID  STATE  NAME\n");
-  printk("---  -----  ----\n");
+int32_t ticks_remaining;
+volatile bool need_resched = false;
 
-  for (int32_t i = 0; i < NUM_PROC; i += 1) {
-    if (ptables[i].state != UNUSED) {
-      const char *state_str[] = {"UNUSED", "EMBRYO",  "SLEEPING",
-                                 "READY",  "RUNNING", "ZOMBIE"};
+/* Private */
 
-      printk("%d  %s  %s\n", ptables[i].pid, state_str[ptables[i].state],
-             ptables[i].name);
-    }
-  }
-}
-
-proc_t *alloc_proc(void) {
+static proc_t *alloc_proc(void) {
   for (int32_t i = 0; i < NUM_PROC; i += 1) {
     if (ptables[i].state == UNUSED) {
       ptables[i].state = EMBRYO;
@@ -52,6 +44,27 @@ void fork_return(void) {
                    :
                    : "r"(retval)
                    : "eax");
+}
+
+/* Public */
+
+inline proc_t *myproc(void) { return current_proc; }
+
+void init_ptables(void) {
+  for (int32_t i = 0; i < NUM_PROC; i += 1) {
+    ptables[i].state = UNUSED;
+  }
+}
+
+inline void check_resched(void) {
+  proc_t *p = myproc();
+  if (!p || need_resched == false) {
+    return;
+  }
+
+  printk("Preempting PID %d\n", current_proc->pid);
+  need_resched = false;
+  yield();
 }
 
 pid_t fork(char *name, void (*f)(void)) {
@@ -84,8 +97,48 @@ pid_t fork(char *name, void (*f)(void)) {
   return p->pid;
 }
 
-void init_ptables(void) {
-  for (int32_t i = 0; i < NUM_PROC; i += 1) {
-    ptables[i].state = UNUSED;
+inline void yield(void) {
+  proc_t *p = myproc();
+  if (!p) {
+    return;
+  }
+
+  need_resched = false;
+  p->state = READY;
+  swtch(&p->context, scheduler_context);
+}
+
+void schedule(void) {
+  static char *scheduler_stack = NULL;
+  if (!scheduler_stack) {
+    scheduler_stack = kmalloc(PAGE_SIZE);
+    if (!scheduler_stack) {
+      abort("Cannot allocate scheduler stack");
+    }
+    scheduler_context = (context_t *)(scheduler_stack + PAGE_SIZE - 64);
+  }
+
+  // FIFO
+  while (true) {
+    for (int32_t i = 0; i < NUM_PROC; i += 1) {
+      if (ptables[i].state != READY) {
+        continue;
+      }
+      proc_t *p = &ptables[i];
+
+      // printk("Scheduler: switching to process %d\n", p->pid);
+
+      p->state = RUNNING;
+      current_proc = p;
+      ticks_remaining = TIME_QUANTUM;
+
+      swtch(&scheduler_context, p->context);
+
+      // printk("Scheduler: back from process %d\n", p->pid);
+      current_proc = NULL;
+    }
+
+    sti();
+    __asm__ volatile("hlt");
   }
 }
